+++ lib/gg_router.dart
@@ -10,5 +10,5 @@ export 'src/gg_router_delegate.dart';
 export 'src/gg_route_information_parser.dart';
 
 export 'src/gg_router.dart';
-export 'src/gg_popover_router.dart';
+export 'src/gg_popover_route.dart';
 export 'src/gg_route_params.dart';